# [refactor] Optimize vector and matrix ndarray fill