--- a/python/taichi/lang/_ndarray.py
+++ b/python/taichi/lang/_ndarray.py
@@ -74,7 +74,8 @@ class Ndarray:
         """
         raise NotImplementedError()
 
-    def ndarray_fill(self, val):
+    @python_scope
+    def fill(self, val):
         """Fills ndarray with a specific scalar value.
 
         Args:
@@ -83,7 +84,7 @@ class Ndarray:
         if impl.current_cfg().ndarray_use_torch:
             self.arr.fill_(val)
         elif impl.current_cfg().arch != _ti_core.Arch.cuda:
-            self.buffer_fill(val)
+            self.fill_by_kernel(val)
         elif self.dtype == primitive_types.f32:
             self.arr.fill_float(val)
         elif self.dtype == primitive_types.i32:
@@ -91,7 +92,7 @@ class Ndarray:
         elif self.dtype == primitive_types.u32:
             self.arr.fill_uint(val)
         else:
-            self.buffer_fill(val)
+            self.fill_by_kernel(val)
 
     def ndarray_to_numpy(self):
         """Converts ndarray to a numpy array.
@@ -253,10 +254,6 @@ class ScalarNdarray(Ndarray):
         return self.host_accessor.getter(*self.pad_key(key))
 
     @python_scope
-    def fill(self, val):
-        self.ndarray_fill(val)
-
-    @python_scope
     def to_numpy(self):
         return self.ndarray_to_numpy()
 
@@ -269,7 +266,7 @@ class ScalarNdarray(Ndarray):
         ret_arr.copy_from(self)
         return ret_arr
 
-    def buffer_fill(self, val):
+    def fill_by_kernel(self, val):
         taichi.lang.meta.fill_ndarray(self, val)
 
     def __repr__(self):