--- a/python/taichi/lang/matrix.py
+++ b/python/taichi/lang/matrix.py
@@ -1349,10 +1349,6 @@ class MatrixNdarray(Ndarray):
              for i in range(self.n)])
 
     @python_scope
-    def fill(self, val):
-        self.ndarray_fill(val)
-
-    @python_scope
     def to_numpy(self):
         return self.ndarray_matrix_to_numpy(as_vector=0)
 
@@ -1366,7 +1362,7 @@ class MatrixNdarray(Ndarray):
         ret_arr.copy_from(self)
         return ret_arr
 
-    def buffer_fill(self, val):
+    def fill_by_kernel(self, val):
         taichi.lang.meta.fill_ndarray_matrix(self, val)
 
     def __repr__(self):
@@ -1409,10 +1405,6 @@ class VectorNdarray(Ndarray):
             [NdarrayHostAccess(self, key, (i, )) for i in range(self.n)])
 
     @python_scope
-    def fill(self, val):
-        self.ndarray_fill(val)
-
-    @python_scope
     def to_numpy(self):
         return self.ndarray_matrix_to_numpy(as_vector=1)
 
@@ -1425,7 +1417,7 @@ class VectorNdarray(Ndarray):
         ret_arr.copy_from(self)
         return ret_arr
 
-    def buffer_fill(self, val):
+    def fill_by_kernel(self, val):
         taichi.lang.meta.fill_ndarray_matrix(self, val)
 
     def __repr__(self):