Speedup ndarray `fill` using `cuMemset32D` and `std::fill` instead of the kernel-based approach. Measured speedups (using the test script from [PR 3907](https://github.com/taichi-dev/taichi/pull/3907):

`ti.Matrix.ndarray(4, 4, ti.f32, N)`

1. `ti.init(arch=ti.cuda) `

| `iterations` | Master | Head |
| ------------- | ------------- | ------------- |
| 10  | 0.457 s| 0.050 ms  |
| 100  | 4.705 s | 0.475 ms |
| 1000  |  65.971 s | 4.374 ms |

2. `ti.init(arch=ti.cpu) `

| `iterations` | Master | Head |
| ------------- | ------------- | ------------- |
| 10  | 0.407 s| 0.150 s  |
| 100  | 4.144 s | 1.488 s |
| 1000  | 42.127 s  | 14.693 s |